# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Learnova API
  version: 1.0.0
  description: |
    API Contract cho hệ thống Learnova.

    Backend: Spring Boot
    Frontend: Next.js

    Lỗi dùng application/problem+json với code ổn định và fieldErrors.
    detail là thông báo an toàn cho người dùng; không có trường message trùng nghĩa.
    F03 bổ sung local authentication, refresh rotation và multi-role.
    F04 bổ sung Google OIDC, onboarding và liên kết Local account có xác nhận.
    F05 bổ sung profile, avatar và đổi mật khẩu Local giữ refresh session hiện tại.
    F06 bổ sung Classroom theo owner, membership và Join Code có preview trước khi join.
    Actuator không được expose qua HTTP. Endpoint chưa khai báo vẫn bị từ chối.
    Các list API dùng page/size và PageResponse; sort được định nghĩa riêng
    theo whitelist của từng endpoint, không nhận property persistence tùy ý.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/attempts/{id}/heartbeat:
    post:
      tags: [Monitoring]
      operationId: heartbeatAttempt
      summary: Record presence for an owned active attempt
      description: |
        ACTIVE PARTICIPANT required. Records server time only while IN_PROGRESS and before persisted deadline.
        No-op after completion/deadline. Does not change answers, revisions, deadline or submission.
        Client sends every 15 seconds while exam taking is mounted. Cache-Control no-store.
      security: [{bearerAuth: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '204': {description: Heartbeat processed.}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404':
          description: ATTEMPT_NOT_FOUND, including another participant's attempt.
          content:
            application/problem+json:
              schema: {$ref: '#/components/schemas/ApiProblem'}
...
  /api/v1/exam-sessions/{id}/results/{participantId}/attempts:
    get:
      tags:
      - Results
      operationId: listCreatorParticipantResults
      summary: Creator owner views all attempts of a participant
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: participantId
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Success. Cache-Control no-store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAttemptPage'
        '400':
          description: Invalid pagination or identifier
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Active business role required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Resource missing or belongs to another user
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '503':
          description: Service unavailable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      description: Participant requires ACTIVE PARTICIPANT and ownership, even after
        classroom removal. HIDDEN never discloses results. AFTER_SESSION_END uses current
        session endTime and server time, including extensions. IMMEDIATE requires GRADED.
        MANUAL requires persisted release. No scores or answers are added to submit/discovery
        responses.
...
  /api/v1/exam-sessions/{id}/attempts:
    post:
      tags: [Attempts]
      operationId: startAttempt
      summary: Start or return the existing IN_PROGRESS Attempt
      description: |
        ACTIVE PARTICIPANT required. Session lock serializes start, configuration and cancellation.
        Existing IN_PROGRESS is checked before assignment. If overdue, it is finalized and returned without consuming another attempt.
        New attempts require current assignment, effective OPEN state and remaining attempts.
        Deadline and concrete question/option order are fixed at creation. All responses use Cache-Control no-store.
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '201':
          description: Created atomically with persisted question and option order.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TakingAttempt'}
        '200':
          description: Existing Attempt; retries do not consume another attempt.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TakingAttempt'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403':
          description: FORBIDDEN, ACCOUNT_LOCKED/ACCOUNT_DISABLED or SESSION_NOT_ASSIGNED.
          content:
            application/problem+json:
              schema: {$ref: '#/components/schemas/ApiProblem'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/AttemptConflict'}
        '503': {$ref: '#/components/responses/AttemptUnavailable'}
  /api/v1/attempts/{id}:
    get:
      tags: [Attempts]
      operationId: readAttempt
      summary: Resume own Attempt with saved state and fixed order
      description: |
        ACTIVE PARTICIPANT owner required; membership is not rechecked. Other owners return 404.
        Only editable attempts include questions. Expired or completed attempts return metadata,
        canEdit=false and questions=[]. Overdue attempts are lazily finalized and graded before response. No result disclosure.
        Cache-Control no-store. serverTime and deadline are authoritative, even when scheduler is late.
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Own Attempt; no correct answers, explanations or grading rules.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TakingAttempt'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/AttemptUnavailable'}
  /api/v1/attempts/{id}/submit:
    post:
      tags: [Attempts]
      operationId: submitAttempt
      summary: Finalize and automatically grade own Attempt exactly once
      description: |
        ACTIVE PARTICIPANT owner required; membership is not rechecked. No request body.
        Client must wait for acknowledged saves before calling. Uses persisted answers and published snapshot.
        Attempt lock serializes autosave, submit and expiration. At or after persisted deadline,
        completionReason is DEADLINE_REACHED; otherwise PARTICIPANT_SUBMIT.
        A successful transaction persists one result and question scores and returns GRADED.
        Retrying returns current metadata without regrading. submittedAt is the accepted server submit time
        or persisted deadline for expiration; gradedAt is processing time. No score, pass/fail or answer key
        is disclosed in F14. Cache-Control no-store. On timeout, read authoritative state before retry.
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Completed Attempt or idempotent current state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TakingAttempt'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/AttemptUnavailable'}
  /api/v1/attempts/{id}/answers/{questionId}:
    put:
      tags: [Attempts]
      operationId: saveAttemptAnswer
      summary: Replace one answer and review state using its server revision
      description: |
        Owner, IN_PROGRESS and now < persisted deadline are checked while holding the Attempt lock.
        Question and option IDs must belong to its published snapshot. Revision is per question.
        Deadline is checked before answer revision. A late save finalizes the attempt without saving its input and returns ATTEMPT_DEADLINE_PASSED after commit.
        Already completed attempts return ATTEMPT_NOT_EDITABLE. Stale editable answers return ANSWER_REVISION_CONFLICT; after a lost response,
        read the Attempt to reconcile before choosing whether to save again. No automatic overwrite.
        activeTimeMs is cumulative estimated focused time, nullable if absent; accepted values are
        capped to elapsed server time and never decrease. It is not grading or anti-cheat evidence.
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: questionId, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SaveAttemptAnswer'}
      responses:
        '200':
          description: Persisted state and incremented revision; Cache-Control no-store.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [state, serverTime]
                properties:
                  state: {$ref: '#/components/schemas/AttemptAnswerState'}
                  serverTime: {type: string, format: date-time}
        '400':
          description: INVALID_ATTEMPT_ANSWER, VALIDATION_FAILED or INVALID_REQUEST; invalid shape, precision or foreign/duplicate option IDs.
          content:
            application/problem+json:
              schema: {$ref: '#/components/schemas/ApiProblem'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404':
          description: ATTEMPT_NOT_FOUND or ATTEMPT_QUESTION_NOT_FOUND, including other owners.
          content:
            application/problem+json:
              schema: {$ref: '#/components/schemas/ApiProblem'}
        '409': {$ref: '#/components/responses/AttemptConflict'}
        '503': {$ref: '#/components/responses/AttemptUnavailable'}
...
  /api/v1/participant/exam-sessions/{id}/attempts:
    get:
      tags:
      - Participant Discovery
      operationId: listParticipantSessionAttempts
      summary: Read own attempt history metadata
      description: Same visibility checks as session detail. Only caller attempts, including
        after membership removal; sorted startedAt descending then id ascending. No
        answers, scores, pass/fail or result link. F13-F15 add write/resume/result flows.
      security:
      - bearerAuth: []
      parameters:
      - *id002
      - *id003
      - *id004
      responses:
        '200':
          description: Participant-safe metadata only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryAttemptPage'
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '503': *id001
...
components:
  securitySchemes:
    googleFlowCookie:
      type: apiKey
      in: cookie
      name: learnova_google_flow
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: learnova_refresh
  parameters:
...
    Page:
      name: page
      in: query
      description: Trang bắt đầu từ 0. Chỉ nhận một số nguyên không âm; sai trả 400.
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 2147483647
        default: 0
    Size:
      name: size
      in: query
      description: Kích thước trang. Rỗng, lặp tham số, không nguyên hoặc ngoài giới hạn trả 400.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    AttemptConflict:
      description: ANSWER_REVISION_CONFLICT, ATTEMPT_DEADLINE_PASSED, ATTEMPT_NOT_EDITABLE, ATTEMPTS_EXHAUSTED or SESSION_INVALID_STATE. No mutation committed.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
    AttemptUnavailable:
      description: SERVICE_UNAVAILABLE; retry with the same revision or read to reconcile an uncertain outcome.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
...
    BadRequest:
      description: INVALID_REQUEST hoặc VALIDATION_FAILED; không trả rejectedValue.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
          example:
            type: about:blank
            title: Bad Request
            status: 400
            detail: The request is invalid.
            instance: /api/v1/questions
            code: VALIDATION_FAILED
            fieldErrors:
              - field: size
                message: Must be between 1 and 100.
    Unauthorized:
      description: AUTHENTICATION_REQUIRED; không redirect đến login.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    Forbidden:
      description: ACCESS_DENIED, bao gồm CSRF không hợp lệ.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    NotFound:
      description: RESOURCE_NOT_FOUND khi request đã qua kiểm tra Security.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
...
  schemas:
...
    ResultScore:
      type: object
      additionalProperties: false
      required:
      - score
      - totalScore
      - passed
      properties:
        score:
          type: string
        totalScore:
          type: string
        passed:
          type: boolean
    ResultAvailability:
      type: string
      enum:
      - AVAILABLE
      - HIDDEN
      - PENDING_RELEASE
      - PENDING_GRADING
    ResultSummary:
      type: object
      additionalProperties: false
      required:
      - correctCount
      - incorrectCount
      - unansweredCount
      - durationSeconds
      properties:
        correctCount:
          type: integer
          minimum: 0
        incorrectCount:
          type: integer
          minimum: 0
        unansweredCount:
          type: integer
          minimum: 0
        durationSeconds:
          type: integer
          minimum: 0
    ResultQuestion:
      type: object
      additionalProperties: false
      required:
      - id
      - type
      - content
      - options
      - answer
      - correctBoolean
      - correctValue
      - tolerance
      - explanation
      - correct
      - points
      - awardedScore
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        content:
          type: string
        options:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
            - id
            - content
            - correct
            properties:
              id:
                type: string
                format: uuid
              content:
                type: string
              correct:
                type: boolean
        answer:
          $ref: '#/components/schemas/ParticipantAnswer'
        correctBoolean:
          type:
          - boolean
          - 'null'
        correctValue:
          type:
          - string
          - 'null'
        tolerance:
          type:
          - string
          - 'null'
        explanation:
          type:
          - string
          - 'null'
        correct:
          type: boolean
        points:
          type: string
        awardedScore:
          type: string
    ResultDetail:
      type: object
      additionalProperties: false
      required:
      - id
      - sessionId
      - title
      - attemptNumber
      - status
      - displayMode
      - availability
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        title:
          type: string
        attemptNumber:
          type: integer
          minimum: 0
        status:
          type: string
        submittedAt:
          type: string
          format: date-time
        displayMode:
          type: string
          enum:
          - HIDDEN
          - SCORE_ONLY
          - SUMMARY
          - DETAILED
        availability:
          $ref: '#/components/schemas/ResultAvailability'
        result:
          $ref: '#/components/schemas/ResultScore'
        summary:
          $ref: '#/components/schemas/ResultSummary'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/ResultQuestion'
      description: 'Unavailable: result, summary and questions are absent. SCORE_ONLY:
        result only. SUMMARY adds summary. DETAILED adds questions from immutable snapshot
        in attempt order. Creator owner sees DETAILED only for GRADED attempts. All scores
        are exact decimal strings.'
...
    ResultAttemptPage:
      allOf:
      - $ref: '#/components/schemas/PageResponse'
      - type: object
        properties:
          content:
            type: array
            items:
              $ref: '#/components/schemas/ResultDetail'
...
    ParticipantAnswer:
      type: object
      additionalProperties: false
      required: [optionIds]
      description: |
        SINGLE_CHOICE accepts zero or one option; MULTIPLE_CHOICE accepts a unique set.
        TRUE_FALSE uses booleanValue (false is an answer). NUMERIC_ANSWER uses numericValue.
        Fields for other types must be null/empty. Clear using empty optionIds and null values.
      properties:
        optionIds: {type: array, maxItems: 20, uniqueItems: true, items: {type: string, format: uuid}}
        booleanValue: {type: [boolean, 'null']}
        numericValue:
          type: [string, 'null']
          maxLength: 42
          pattern: '^-?\d+(\.\d+)?$'
          description: Decimal string, no exponent; at most 20 integer digits and 10 fractional digits.
    SaveAttemptAnswer:
      type: object
      additionalProperties: false
      required: [revision, answer, markedForReview]
      properties:
        revision: {type: integer, format: int64, minimum: 0}
        answer: {$ref: '#/components/schemas/ParticipantAnswer'}
        markedForReview: {type: boolean}
        activeTimeMs: {type: [integer, 'null'], format: int64, minimum: 0}
    AttemptAnswerState:
      type: object
      additionalProperties: false
      required: [questionId, answer, markedForReview, revision, activeTimeMs, savedAt]
      properties:
        questionId: {type: string, format: uuid}
        answer: {$ref: '#/components/schemas/ParticipantAnswer'}
        markedForReview: {type: boolean}
        revision: {type: integer, format: int64, minimum: 0}
        activeTimeMs: {type: [integer, 'null'], format: int64, minimum: 0}
        savedAt: {type: [string, 'null'], format: date-time}
    TakingQuestion:
      type: object
      additionalProperties: false
      required: [id, type, content, options, state]
      properties:
        id: {type: string, format: uuid}
        type: {type: string, enum: [SINGLE_CHOICE, MULTIPLE_CHOICE, TRUE_FALSE, NUMERIC_ANSWER]}
        content: {type: string}
        options:
          type: array
          description: Persisted order; empty for TRUE_FALSE and NUMERIC_ANSWER. Labels are assigned from this display order.
          items:
            type: object
            additionalProperties: false
            required: [id, content]
            properties:
              id: {type: string, format: uuid}
              content: {type: string}
        state: {$ref: '#/components/schemas/AttemptAnswerState'}
    TakingAttempt:
      type: object
      additionalProperties: false
      required: [id, sessionId, examVersionId, title, attemptNumber, status, startedAt, deadline, serverTime, canEdit, questions, completionReason, submittedAt, gradedAt]
      properties:
        id: {type: string, format: uuid}
        sessionId: {type: string, format: uuid}
        examVersionId: {type: string, format: uuid}
        title: {type: string}
        attemptNumber: {type: integer, minimum: 1}
        status: {type: string, enum: [IN_PROGRESS, SUBMITTED, EXPIRED, GRADED]}
        startedAt: {type: string, format: date-time}
        deadline: {type: string, format: date-time}
        serverTime: {type: string, format: date-time}
        completionReason:
          type: [string, 'null']
          enum: [PARTICIPANT_SUBMIT, DEADLINE_REACHED, null]
        submittedAt: {type: [string, 'null'], format: date-time}
        gradedAt: {type: [string, 'null'], format: date-time}
        canEdit: {type: boolean}
        questions:
          type: array
          items: {$ref: '#/components/schemas/TakingQuestion'}
...
    DiscoveryAttempt:
      type: object
      required:
      - id
      - attemptNumber
      - status
      - startedAt
      - deadline
      - submittedAt
      properties:
        id:
          type: string
          format: uuid
        attemptNumber:
          type: integer
          minimum: 1
        status:
          type: string
          enum:
          - IN_PROGRESS
          - SUBMITTED
          - EXPIRED
          - GRADED
        startedAt:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
        submittedAt:
          type:
          - string
          - 'null'
          format: date-time
...
    DiscoveryAttemptPage:
      type: object
      required:
      - content
      - page
      - size
      - totalElements
      - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveryAttempt'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 0
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    ApiProblem:
      type: object
      required: [type, title, status, detail, instance, code, fieldErrors]
      properties:
        type:
          type: string
          format: uri-reference
          example: about:blank
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
          description: Path của request, không chứa query string.
        code:
          type: string
          description: |
            Mã ổn định. F01: INVALID_REQUEST/VALIDATION_FAILED (400),
            AUTHENTICATION_REQUIRED (401), ACCESS_DENIED (403), RESOURCE_NOT_FOUND (404),
            METHOD_NOT_ALLOWED (405), NOT_ACCEPTABLE (406), CONFLICT (409),
            PAYLOAD_TOO_LARGE (413), UNSUPPORTED_MEDIA_TYPE (415), INTERNAL_ERROR (500),
            SERVICE_UNAVAILABLE (503); fallback REQUEST_REJECTED cho lỗi 4xx còn lại.
            Feature sau có thể bổ sung business code.
        fieldErrors:
          type: array
          description: Danh sách rỗng nếu không có lỗi theo field; không chứa giá trị đầu vào.
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    PageResponse:
      type: object
      description: |
        Envelope dùng chung; từng list operation phải khai báo schema item cụ thể
        trong content. Trang vượt tổng số trang trả content rỗng và giữ page yêu cầu.
        Kết quả rỗng có totalElements=0, totalPages=0. Không serialize Spring Page.
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items: {}
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...